add e2e cases about preemption
